+++ src/js/pages/demo/bmrichtext/index.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script>
-if (process.env.NODE_ENV === 'development') require('Config');
+
 export default {
     data () {
         return {};