+++ CHANGELOG.md
@@ -1,8 +1,11 @@
-## date: 2018-02-01
+## 2018-03-05
+* 模板更新： 删除所有if (process.env.NODE_ENV === 'development') require('Config');
+* 模板更新： 添加 ios 项目 schemes，platforms/ios/WeexEros/WeexEros.xcodeproj/xcshareddata/
+## 2018-02-01
 * 模板更新： 全新 demo
 * 模板更新： 添加 scripts 文件夹 内置脚本
 
-## date: 2017-12-29
+## 2017-12-29
 * 模板更新： widget 更新优化
 * 模板更新： 移除__eros__文件夹 不再需要
 * 模板更新： 支持 npm install weex-ui 来使用